Conflicts:
	.gitignore
	src/mongo/db/audit.cpp

+++ src/mongo/db/storage/ephemeral_for_test/ephemeral_for_test_record_store.cpp
@@ -291,7 +291,7 @@ void RecordStore::appendCustomStats(OperationContext* opCtx,
     result->appendBool("capped", _isCapped);
     if (_isCapped) {
         result->appendNumber("max", static_cast<long long>(_cappedMaxDocs));
-        result->appendNumber("maxSize", _cappedMaxSize / scale);
+        result->appendNumber("maxSize", static_cast<long long>(_cappedMaxSize / scale));
     }
 }
